+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ##### Enhancements
 
+* Added new syntax, attribute and declaration kinds introduced in Swift 5.9.  
+  [John Fairhurst](https://github.com/johnfairh)
+
 * Added new syntax, attribute and declaration kinds introduced in Swift 5.8.  
   [JP Simard](https://github.com/jpsim)